Move each subcommand to its own mod.
