+++ probe-rs/src/bin/probe-rs/gdb.rs
@@ -3,39 +3,58 @@ use std::time::Duration;
 
 use crate::util::common_options::ProbeOptions;
 
-pub fn run_gdb_server(
-    common: ProbeOptions,
-    connection_string: Option<&str>,
+#[derive(clap::Parser)]
+pub struct Cmd {
+    #[clap(
+        long,
+        help = "Use this flag to override the default GDB connection string (localhost:1337)."
+    )]
+    gdb_connection_string: Option<String>,
+
+    #[clap(
+        name = "reset-halt",
+        long = "reset-halt",
+        help = "Use this flag to reset and halt (instead of just a halt) the attached core after attaching to the target."
+    )]
     reset_halt: bool,
-) -> anyhow::Result<()> {
-    let mut session = common.simple_attach()?;
 
-    if reset_halt {
-        session
-            .core(0)?
-            .reset_and_halt(Duration::from_millis(100))?;
-    }
+    #[clap(flatten)]
+    common: ProbeOptions,
+}
 
-    let gdb_connection_string = connection_string.unwrap_or("localhost:1337");
+impl Cmd {
+    pub fn run(self) -> anyhow::Result<()> {
+        let mut session = self.common.simple_attach()?;
 
-    let instances = probe_rs::gdb_server::GdbInstanceConfiguration::from_session(
-        &session,
-        Some(gdb_connection_string.to_owned()),
-    );
+        if self.reset_halt {
+            session
+                .core(0)?
+                .reset_and_halt(Duration::from_millis(100))?;
+        }
 
-    for instance in instances.iter() {
-        println!(
-            "Firing up GDB stub for {:?} cores at {:?}",
-            instance.core_type, instance.socket_addrs
+        let gdb_connection_string = self
+            .gdb_connection_string
+            .unwrap_or_else(|| "localhost:1337".to_string());
+
+        let instances = probe_rs::gdb_server::GdbInstanceConfiguration::from_session(
+            &session,
+            Some(gdb_connection_string),
         );
+
+        for instance in instances.iter() {
+            println!(
+                "Firing up GDB stub for {:?} cores at {:?}",
+                instance.core_type, instance.socket_addrs
+            );
+        }
+
+        let session = Mutex::new(session);
+
+        if let Err(e) = probe_rs::gdb_server::run(&session, instances.iter()) {
+            eprintln!("During the execution of GDB an error was encountered:");
+            eprintln!("{e:?}");
+        }
+
+        Ok(())
     }
-
-    let session = Mutex::new(session);
-
-    if let Err(e) = probe_rs::gdb_server::run(&session, instances.iter()) {
-        eprintln!("During the execution of GDB an error was encountered:");
-        eprintln!("{e:?}");
-    }
-
-    Ok(())
 }
